refactor(build-manager): simplify build runner controller and moved logic and kafka emmit to the service
