+++ packages/amplication-build-manager/src/redis/redis.service.ts
@@ -25,6 +25,10 @@ export class RedisService {
       : this.redisClient.set(key, serializedValue);
   }
 
+  async delete(key: string): Promise<number> {
+    return this.redisClient.del(key);
+  }
+
   private serializeValue<T>(value: T): string {
     return JSON.stringify(value);
   }
